Moved some old classes to a legacy package
