+++ src/me/simonhaasnoot/geinigspel/legacy/Character.java
@@ -1,8 +1,6 @@
-package me.simonhaasnoot.geinigspel.Entiteiten;
+package me.simonhaasnoot.geinigspel.legacy;
 
 import me.simonhaasnoot.geinigspel.Values;
-
-import java.awt.*;
 
 /**
  * Created by Simon on 2/18/2016.
