+++ src/me/simonhaasnoot/geinigspel/legacy/Meteorite.java
@@ -1,7 +1,5 @@
-package me.simonhaasnoot.geinigspel.Entiteiten;
+package me.simonhaasnoot.geinigspel.legacy;
 import me.simonhaasnoot.geinigspel.Values;
-
-import java.awt.*;
 
 /**
  * Created by Simon on 2/18/2016.
